pyproject.toml
[project]
name = "nash-api"
version = "2.0.0"
description = "Nash 2.0 Python API backend"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1",
    "flask-cors>=5.0",
    "backboard-sdk",
    "pyjwt[crypto]>=2.9",
    "authlib>=1.3",
    "httpx>=0.28",
    "stripe>=11.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "gunicorn>=23.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = []
